Move a unused but potentially useful helper function in utils
