+++ vivarium/environment/components/eco_evo/consumption/component.py
@@ -1,4 +1,3 @@
-import jax
 import jax.numpy as jnp
 from jax_md import partition
 from jax_md.dataclasses import dataclass as md_dataclass
@@ -15,34 +14,7 @@ class ConsumptionState:
     target_subtype: jnp.ndarray
     range: jnp.ndarray
     start: jnp.ndarray
-
-
-def count_masked_values(x, mask, num_values):
-    """
-    For each integer value in x, count how many times it corresponds 
-    to True values in mask.
     
-    Args:
-        x: array of integers, shape (N,), with values in [0, num_values)
-        mask: array of booleans, shape (N,)
-        num_values: the range of possible values [0, num_values).
-                   Must be a concrete integer (not a traced value).
-    
-    Returns:
-        counts: array of shape (num_values,) where counts[i] is the number
-                of times value i appears in x where mask is True
-    """
-    # Use segment_sum to count occurrences
-    # Convert mask to integers (True -> 1, False -> 0)
-    counts = jax.ops.segment_sum(
-        mask.astype(jnp.int32),
-        x,
-        num_segments=num_values
-    )
-    
-    return counts
-
-count_masked_values = jax.jit(count_masked_values, static_argnums=(2,))
 
 class ConsumptionComponent(Component):
     def __init__(self, name, precedence, source_subtype, target_subtype, range, start):
