+++ vivarium/environment/components/utils.py
@@ -1,5 +1,6 @@
 import logging
 
+import jax
 import jax.numpy as jnp
 
 from jax_md import rigid_body, util
@@ -36,3 +37,31 @@ def sum_force_fns(displacement, force_fns):
         force = sum_forces([fn(state, neighbor, exists_mask) for fn in fns])
         return force
     return force_fn
+
+
+def count_masked_values(x, mask, num_values):
+    """
+    For each integer value in x, count how many times it corresponds 
+    to True values in mask.
+    
+    Args:
+        x: array of integers, shape (N,), with values in [0, num_values)
+        mask: array of booleans, shape (N,)
+        num_values: the range of possible values [0, num_values).
+                   Must be a concrete integer (not a traced value).
+    
+    Returns:
+        counts: array of shape (num_values,) where counts[i] is the number
+                of times value i appears in x where mask is True
+    """
+    # Use segment_sum to count occurrences
+    # Convert mask to integers (True -> 1, False -> 0)
+    counts = jax.ops.segment_sum(
+        mask.astype(jnp.int32),
+        x,
+        num_segments=num_values
+    )
+    
+    return counts
+
+count_masked_values = jax.jit(count_masked_values, static_argnums=(2,))
